BREAKING: Use floating point values for velocity Ready for hi-res MIDI, and also helps for GUI stuff down the line

+++ src/BeatDescription.h
@@ -14,7 +14,7 @@ struct Note {
     double timestamp;
     double duration;
     uint8_t number;
-    uint8_t velocity;
+    float velocity;
 };
 
 using Sequence = std::vector<Note>;
